fork-main/requirements.txt
# NodeNAR 项目依赖项
# 基于用户当前环境生成的 requirements.txt

# 核心深度学习框架
torch==2.1.0+cu121
torchvision==0.16.0+cu121
torchaudio==2.1.0+cu121
torch-geometric==2.5.2
torchmetrics==1.5.2

# 图神经网络相关
dive-into-graphs==1.1.0
dig==0.1.10
networkx==3.0

# 化学信息学
rdkit==2024.3.2
rdkit-pypi==2022.9.5
pyscf==2.7.0

# 科学计算
numpy==1.24.3
scipy==1.10.1
scikit-learn==1.3.0
pandas==2.0.3

# 可视化和绘图
matplotlib==3.7.5
seaborn==0.13.2
plotly==5.17.0

# 配置和工具
PyYAML==6.0.2
hydra-core==1.3.2
omegaconf==2.3.0
easydict==1.13

# 进度条和日志
tqdm==4.66.5
wandb==0.19.0

# 其他工具库
einops==0.8.0
einops-exts==0.0.4
einops-layers==0.4.0

# 开发和调试工具
ipdb==0.13.13
jupyter_client==8.6.3
ipykernel==6.29.5

# 文件处理
h5py==3.11.0
safetensors==0.5.2

# 网络请求
requests==2.32.3
aiohttp==3.10.5

# 其他依赖
attrs==24.2.0
click==8.1.7
decorator==5.1.1
joblib==1.4.2
packaging==24.1
pillow==10.4.0
protobuf==5.29.1
psutil==5.9.0
pydantic==2.10.3
pydantic_core==2.27.1
python-dateutil==2.9.0.post0
pytz==2024.2
setuptools==75.3.0
six==1.16.0
sympy==1.13.1
tabulate==0.9.0
threadpoolctl==3.5.0
toolz==1.0.0
typing_extensions==4.12.2
urllib3==1.26.13
zipp==3.20.2 